Move most of App component to separate Dashboard component
